+++ src/main/java/com/souf/soufwebsite/domain/member/controller/admin/AdminController.java
@@ -1,12 +1,12 @@
 package com.souf.soufwebsite.domain.member.controller.admin;
 
-import com.souf.soufwebsite.domain.file.entity.PostType;
 import com.souf.soufwebsite.domain.member.dto.ResDto.AdminMemberResDto;
 import com.souf.soufwebsite.domain.member.dto.ResDto.AdminPostResDto;
 import com.souf.soufwebsite.domain.member.dto.ResDto.AdminReportResDto;
 import com.souf.soufwebsite.domain.member.entity.RoleType;
 import com.souf.soufwebsite.domain.member.service.admin.AdminService;
 import com.souf.soufwebsite.domain.report.entity.ReportStatus;
+import com.souf.soufwebsite.global.common.PostType;
 import com.souf.soufwebsite.global.success.SuccessResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
