+++ src/main/java/com/souf/soufwebsite/domain/member/dto/ResDto/AdminReportResDto.java
@@ -1,7 +1,7 @@
 package com.souf.soufwebsite.domain.member.dto.ResDto;
 
-import com.souf.soufwebsite.domain.file.entity.PostType;
 import com.souf.soufwebsite.domain.report.entity.ReportStatus;
+import com.souf.soufwebsite.global.common.PostType;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDateTime;
