+++ src/main/java/com/souf/soufwebsite/domain/report/repository/ReportCustomRepository.java
@@ -1,7 +1,7 @@
 package com.souf.soufwebsite.domain.report.repository;
 
-import com.souf.soufwebsite.domain.file.entity.PostType;
 import com.souf.soufwebsite.domain.member.dto.ResDto.AdminReportResDto;
+import com.souf.soufwebsite.global.common.PostType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
