+++ src/main/java/com/souf/soufwebsite/domain/report/repository/ReportCustomRepositoryImpl.java
@@ -4,12 +4,12 @@ import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.Tuple;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.souf.soufwebsite.domain.file.entity.PostType;
 import com.souf.soufwebsite.domain.member.dto.ResDto.AdminReportResDto;
 import com.souf.soufwebsite.domain.member.entity.QMember;
 import com.souf.soufwebsite.domain.report.entity.QReason;
 import com.souf.soufwebsite.domain.report.entity.QReportReasonMapping;
 import com.souf.soufwebsite.domain.report.entity.ReportStatus;
+import com.souf.soufwebsite.global.common.PostType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
